refactor malware scanning
Extract shared components into MalwareScanning module
and include it in the new controller and the base_file_uploader_form

Refactor the error_path handling to allow it to be identical from
each calling class
